finish  2 fibo test done

+++ src/VMtranslator.java
@@ -12,6 +12,8 @@ public class VMtranslator {
      * @param args -fles to translate
      */
     public static void main (String[] args){
+        int i =0;
+
         if (args.length != 1) {
             System.out.println("there is something in the inputs arguments");
             return;
@@ -19,7 +21,7 @@ public class VMtranslator {
         try {
 
             File fileIn = new File(args[0]);
-            if (fileIn.getName().endsWith(".vm")) {
+            if (fileIn.isFile() && fileIn.getName().endsWith(".vm")) {
 
                 CodeWriter codewriter = new CodeWriter(fileIn.getName());
                 CodeWriter.outputFile = new BufferedWriter(new FileWriter(fileIn.getCanonicalPath().substring
@@ -50,6 +52,12 @@ public class VMtranslator {
                     if(parser.commandType().equals(Parser.C_CALL)){
                         codewriter.writeFuncCall(parser.firstArg(), parser.seconrArg());
                     }
+                    if(parser.commandType().equals(Parser.C_FUNCTION)){
+                        codewriter.declareFunc(parser.firstArg(), parser.seconrArg());
+                    }
+                    if(parser.commandType().equals(Parser.C_RETURN)){
+                        codewriter.funReturn();
+                    }
 
                 }
                 CodeWriter.outputFile.close();
@@ -59,8 +67,12 @@ public class VMtranslator {
                 CodeWriter codewriter = new CodeWriter(fileIn.getName());
                 CodeWriter.outputFile = new BufferedWriter(new FileWriter(fileIn.getAbsolutePath() +
                                                                         "/" + codewriter.getFileOutPath() + ".asm"));
+                codewriter.writeInit();
+
                 for (File f : files) {
                     if (f.getName().endsWith(".vm")) {
+//                                            System.out.println(f.getName());
+
                         Parser parser = new Parser(f.getName());
                         Parser.reader = new BufferedReader(new FileReader(f.getAbsolutePath()));
                         while (parser.hasMoreCommands()) {
@@ -71,8 +83,28 @@ public class VMtranslator {
                             if (parser.commandType().equals(Parser.C_PUSH) || parser.commandType().
                                                                                                 equals(Parser.C_POP)) {
                                 codewriter.writePushPop(parser.commandType(), parser.firstArg(), parser.seconrArg());
-                            } else if (parser.commandType().equals(Parser.C_ARITHMETIC)) {
+                            } if (parser.commandType().equals(Parser.C_ARITHMETIC)) {
                                 codewriter.writeArithmetic(parser.firstArg());
+                            }
+                            if(parser.commandType().equals(Parser.C_LABEL)){
+                                codewriter.writeLabel(parser.firstArg());
+                            }
+                            if(parser.commandType().equals(Parser.C_GOTO)){
+                                codewriter.writeGoTo(parser.firstArg());
+                            }
+                            if(parser.commandType().equals(Parser.C_IF_GOTO)){
+                                System.out.println(parser.firstArg());
+
+                                codewriter.writeIfGoTo(parser.firstArg());
+                            }
+                            if(parser.commandType().equals(Parser.C_CALL)){
+                                codewriter.writeFuncCall(parser.firstArg(), parser.seconrArg());
+                            }
+                            if(parser.commandType().equals(Parser.C_FUNCTION)){
+                                codewriter.declareFunc(parser.firstArg(), parser.seconrArg());
+                            }
+                            if(parser.commandType().equals(Parser.C_RETURN)){
+                                codewriter.funReturn();
                             }
                         }
                     }
